+++ layers/GraphQLAPIForWP/plugins/graphql-api-for-wp/src/Hooks/VarsHooks.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace GraphQLAPI\GraphQLAPI\Hooks;
 
-use GraphQLAPI\GraphQLAPI\Facades\ModuleRegistryFacade;
+use GraphQLAPI\GraphQLAPI\SystemFacades\ModuleRegistryFacade;
 use GraphQLAPI\GraphQLAPI\SystemServices\ModuleResolvers\EndpointFunctionalityModuleResolver;
 use PoP\API\Response\Schemes as APISchemes;
 use PoP\ComponentModel\Facades\Instances\InstanceManagerFacade;
